--- my-app/src/Contact.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-
-const Contact = () => {
-    return (
-        <>
-        <h1>Hello Contact</h1>
-        </>
-    )
-}
-export default Contact;
