+++ my-app/src/Pages/About.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import web from '../src/images/img-1.png';
-import Comman from './Comman';
+import web from '../assets/images/img-1.png';
+import Comman from '../Pages/Comman';
 
 const About = () => {
     return (
