+++ my-app/src/Pages/Home.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
-import web from '../src/images/img-1.png';
-import Comman from './Comman';
+import web from '../assets/images/img-1.png';
+import Comman from '../Pages/Comman';
 
 const Home = () => {
     return (
