+++ my-app/src/Parts/Footer.jsx
@@ -3,7 +3,7 @@ import React from 'react'
 const Footer = () => {
     return (
         <>
-        <footer className="bg-light text-center">
+        <footer className="bg-black text-center">
             <p>Privcy Policy Terams and Condition privacy</p> 
         </footer>
         </>
