+++ my-app/src/assets/images/img-1.jpg
