+++ my-app/src/assets/style/App.css
@@ -67,3 +67,18 @@ body {
   0% { transform: translateY(0); }
   100% { transform: translateY(-10px); }
 }
+
+.bg-black {
+  background: #000;
+  padding: 5px 0;
+}
+.bg-black p {
+  color: #fff;
+  margin: 0;
+}
+footer {
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  right: 0;
+}
