--- my-app/src/index.css
@@ -1,8 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Mulish:wght@200&display=swap');
-
-body {
-  margin: 0;
-  font-family: 'Mulish', sans-serif;
-  font-style: italic;
-}
-
