+++ my-app/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import App from './App'
+import App from './Pages/App'
 import {BrowserRouter} from 'react-router-dom'
